css/css-position/sticky/position-sticky-scrolled-remove-horiz-sibling.html

<!DOCTYPE html>

  <meta charset="UTF-8">

  <title>CSS Positioned Layout Test: element with 'position: sticky' and removing a sibling in the horizontal axis</title>

  <link rel="author" title="Gérard Talbot" href="http://www.gtalbot.org/BrowserBugsSection/css21testsuite/">

  <!--

  This test is the horizontal-axis correspondent
  to the test
  http://wpt.live/css/css-position/sticky/position-sticky-scrolled-remove-sibling.html
  created by Morten Stenshorne

  Credit given to him here.

  -->

  <link rel="help" href="https://www.w3.org/TR/css-position-3/#sticky-pos">

  <style>
  div#scrollingContainer
    {
      background-color: red;
      height: 100px;
      overflow-x: auto;
      overflow-y: hidden;
      width: 100px;
    }

  div#scrollingContainer > div
    {
      display: inline-block;
      height: 100%;
    }

  div#elemSticky
    {
      background-color: green;
      left: 0px;
      position: sticky;
      width: 100px;
    }

  div#wideItem
    {
      width: 600px;
    }
  </style>

  <script src="/resources/testharness.js"></script>

  <script src="/resources/testharnessreport.js"></script>

  <p>Test passes if there is a filled green square and no horizontal scrollbar.

  <div id="scrollingContainer">

    <div id="elemSticky"></div><div id="wideItem"></div>

  </div>

  <script>
  scrollingContainer.scrollLeft = 600;
    requestAnimationFrame(()=>{
      requestAnimationFrame(()=>{
        wideItem.style.display = "none";
        test(()=> {
        assert_equals(scrollingContainer.scrollWidth, 100);
      }, "Sticky position and its overflow contribution in the horizontal axis");
    });
  });
  </script>
